+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## 2023-12-18
 
+### Added
+
+- Added import save file feature to the Item Tracker. This will allow you to import your save file and see what items you have and what you are missing.
+  - **NOTE:** This is still a work in progress. It is currently finding about 90% of items. I will continue to work on this and improve it ASAP.
+
 ### Fixed
 
 - When selecting mutators, the available mutators are now limited to the corresponding weapon type.